Move Application into its own models_app.py
Separate the application layer (Application plus its application-level
support types: ApplicationParameter, BoundaryType, CompartmentMapping,
SpeciesMapping, ReactionMapping, StructureMapping, Simulation) into
models_app.py. The shared leaf types Parameter and Version move to
models_base.py.

models_app has no runtime dependency on models.py: the model-level types
(Species/Compartment/Reaction) appear only in map_* signatures, imported
under TYPE_CHECKING, with the runtime checks testing for str instead. This
keeps a clean one-directional layering (models -> models_app/base/geometry/
math) with no circular import. models.py re-exports the application types so
the public API is unchanged (vc.Application, from models import Simulation,
etc. all still resolve).

No behavior change. make check passes; tests green.
